requirements.txt
git+https://github.com/azavea/grout.git@develop#egg=ashlar
django-cors-headers==2.2.0
